+++ zeldiablo/src/gameZeldiablo/Zeldiablo/StrategieDeplacement/DeplacementRapprochement.java
@@ -20,18 +20,18 @@ public class DeplacementRapprochement implements DeplacementStrategie, Serializa
     @Override
     public void deplacement(Player joueur, Monstre monstre) {
         // Récupère la position du joueur
-        int joueurX = joueur.getX();
-        int joueurY = joueur.getY();
+        double joueurX = joueur.getX();
+        double joueurY = joueur.getY();
         
-        int monstreX = monstre.getX();
-        int monstreY = monstre.getY();
+        double monstreX = monstre.getX();
+        double monstreY = monstre.getY();
         
         Direction direction = null;
         
         // Récupère la différence entre la position du monstre et celle du joueur
         // pour savoir dans quelle direction se déplacer
-        int deltaX = joueurX - monstreX;
-        int deltaY = joueurY - monstreY;
+        double deltaX = joueurX - monstreX;
+        double deltaY = joueurY - monstreY;
         
         // Regarde là où il y a le plus de différence
         if (Math.abs(deltaX) > Math.abs(deltaY)) {
@@ -41,7 +41,7 @@ public class DeplacementRapprochement implements DeplacementStrategie, Serializa
         }
         
         // Tente le déplacement
-        int[] prochainePosition = Labyrinthe.getSuivant(monstreY, monstreX, direction);
+        int[] prochainePosition = Labyrinthe.getSuivant((int)monstreY, (int)monstreX, direction);
         if (joueur.getLabyrinthe().canEntityMoveTo(prochainePosition[0], prochainePosition[1])) {
             monstre.setPosition(prochainePosition[0], prochainePosition[1]);
         }
